check for existing cmake file in specified test dir

+++ entrypoint.sh
@@ -9,14 +9,28 @@ IFS=';' read -ra paths <<< "$1"
 for i in "${paths[@]}"; do
   echo "processing unit tests in: $i"
   
-  cp /CMakeLists.txt "$i"
+  # Check if CMakeLists.txt already exists in the unit test directory
+  if [ ! -e "$i/CMakeLists.txt" ]; then
+    cp /CMakeLists.txt "$i"
+  fi
+
   pushd "$i" || exit 1
   cmake CMakeLists.txt
   make || exit 2
 
+  # Find all executables ending in _tests recursively in the specified root directory
+  test_executables=($(find "$i" -type f -executable -name "*_tests"))
+
+  # Check if any test executables were found
+  if [ ${#test_executables[@]} -eq 0 ]; then
+    exit 4
+  fi
+
   # execute the unit tests
   popd || exit 3
-  valgrind --leak-check=full --error-exitcode=4 "./$i/unit_tests" || exit 4
+  for executable in "${test_executables[@]}"; do
+    valgrind --leak-check=full --error-exitcode=4 "$executable" || exit 4
+  done
 
   # generate a coverage report
   uuid=$(uuidgen)
